Linux-next for the Ouya Console, Release Alpha 0.10. 31 May 2018
This is the first release of the Linux-next kernel for the Ouya.
**** DANGER! DRAGONS AHEAD! ****
This has been tested on one device only (mine).
It may cause your device to melt or become sentient.
I hold no warranty or liability should you decide to use this!

It supports the Kickstarter Edition Ouya only!
It may work on the newer variant of the Ouya, but I don't have one.

The following changes were made to the source code:

*Trusted Foundations support for the L2 cache controller.
*Cpufreq support for the T30 series.
(This has been merged into the T20 driver in mainline.
*GPT sector forced searching for eMMC support.

Included is the Linux-Next-Ouya-Patch, which is all changes made to the
source, including device tree files.

The following limitations are known:

*Kernel is unstable, it has been known to panic.
*eMMC may be unstable.
*Bluetooth is enabled, but loads with a random MAC address.
*Video support is hit or miss, and there is very little graphics
acceleration.
*Ethernet chip loads with a random MAC address.
*There is no power management to speak of.
*Power off does not work, but reboot does.

The following does work:
*Wifi
*Bluetooth
*eMMC
*Cpufreq
*RTC
*USB
*Ethernet
*Video