ARM: tegra: Remove cpuidle drivers to replace them with a new driver
Remove the old drivers to replace them cleanly with a new one later on.
Please note that old Tegra20 CPUIDLE driver used pen-locking in order to
block secondary CPU waking-up if IRQ happened to raise during of the
entering into CC6 and primary CPU already crossed point of no return, but
that is unnecessary in practice because CPUIDLE simply won't enter the
coupled CC6 state if there is some CPU activity going on. The new driver
will ensure that all secondary CPUs parked successfully at first and thus
primary CPU won't be racing with the secondaries, hence the pen-locking
functionality isn't needed anymore at all. This is the most notable thing
of this patch.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>