ARM: tegra: Add device-tree for ASUS Google Nexus 7
There are two variants of the NVIDIA Tegra30-based Nexus 7 device:
WiFi-only (named Grouper) and GSM (named Tilapia), this patch adds
device-trees for the both variants. The device-trees are mostly
complete, although there are few things that are pending to be
upstreamed yet.

Current status of the supported features could be found here:

https://wiki.postmarketos.org/wiki/Google_Nexus_7_2012_(asus-grouper)#Hardware_status_.28mainline.29

Signed-off-by: David Heidelberg <david@ixit.cz>
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>